pyproject.toml
[project]
name = "windows-mcp"
version = "0.6.9"
description = "Lightweight MCP Server for interacting with Windows Operating System."
authors = [
    { name = "Jeomon George", email = "jeogeoalukka@gmail.com" }
]
readme = "README.md"
license = { file = "LICENSE.md" }
keywords = ["windows", "mcp", "ai", "desktop", "ai agent"]
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Programming Language :: Python :: 3.13",
    "Topic :: Desktop Environment",
    "Topic :: System :: Shells",
]
dependencies = [
    "click>=8.2.1",
    "comtypes>=1.4.15",
    "fastmcp>=3.0",
    "fuzzywuzzy>=0.18.0",
    "ipykernel>=7.2.0",
    "markdownify>=1.1.0",
    "pillow>=11.2.1",
    "platformdirs>=4.3.8",
    "posthog>=7.4.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
    "python-levenshtein>=0.27.1",
    "pywin32>=311",
    "requests>=2.32.3",
    "tabulate>=0.9.0",
    "thefuzz>=0.22.1",
    "uuid7>=0.1.0",
    "winrt-runtime>=3.2.1",
    "winrt-windows-foundation>=3.2.1",
    "winrt-windows-foundation-collections>=3.2.1",
    "winrt-windows-globalization>=3.2.1",
    "winrt-windows-graphics-imaging>=3.2.1",
    "winrt-windows-media-ocr>=3.2.1",
    "winrt-windows-security-cryptography>=3.2.1",
    "winrt-windows-storage>=3.2.1",
    "winrt-windows-storage-streams>=3.2.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/CursorTouch"
Repository = "https://github.com/CursorTouch/Windows-MCP"
Issues = "https://github.com/CursorTouch/Windows-MCP/issues"

[project.scripts]
windows-mcp = "windows_mcp.__main__:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"src/windows_mcp/uia/*.py" = ["F403", "F405"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
